split headers to sources
