+++ tools/EtherBuilder/CMakeLists.txt
@@ -10,13 +10,24 @@ set(CMAKE_CXX_STANDARD 11)
 link_directories(
         ../parser/build
 )
-include_directories(include ../parser/include)
+include_directories(
+        include
+        ../parser/include
+)
+set(SOURCES
+        src/block_store.cpp
+)
+set(HEADERS
+        include/block_store.h
+        include/test_sqlite.h
+)
+
 
 # create the executable
-add_executable(${PROJECT_NAME} main.cpp)
+add_executable(${PROJECT_NAME} main.cpp ${SOURCES} ${HEADERS})
 
 target_link_libraries (${PROJECT_NAME} rocksdb leveldb cryptopp secp256k1 gtest ethereum_parser sqlite3 ${BASE_LIBS})
 
 # researching related
-add_executable(test_sqlite test_sqlite.cpp)
+add_executable(test_sqlite test_sqlite.cpp ${SOURCES} ${HEADERS})
 target_link_libraries (test_sqlite rocksdb leveldb cryptopp secp256k1 gtest ethereum_parser sqlite3 ${BASE_LIBS})
